+++ plugins/Mill/src/main/kotlin/dev/ghostflyby/mill/sdk/MillModuleJdkResolver.kt
@@ -22,10 +22,10 @@
 
 package dev.ghostflyby.mill.sdk
 
-import dev.ghostflyby.mill.MillExecutionSettings
 import dev.ghostflyby.mill.MillImportDebugLogger
 import dev.ghostflyby.mill.command.MillCommandLineUtil
 import dev.ghostflyby.mill.project.MillDiscoveredModule
+import dev.ghostflyby.mill.settings.MillExecutionSettings
 
 internal object MillModuleJdkResolver {
     fun resolve(