+++ plugins/Mill/src/main/kotlin/dev/ghostflyby/mill/sdk/MillScalaSdkResolver.kt
@@ -25,11 +25,11 @@ package dev.ghostflyby.mill.sdk
 import com.intellij.execution.process.ProcessOutputType
 import com.intellij.openapi.externalSystem.model.task.ExternalSystemTaskId
 import com.intellij.openapi.externalSystem.model.task.ExternalSystemTaskNotificationListener
-import dev.ghostflyby.mill.MillExecutionSettings
 import dev.ghostflyby.mill.MillImportDebugLogger
 import dev.ghostflyby.mill.MillScalaSdkData
 import dev.ghostflyby.mill.command.MillCommandLineUtil
 import dev.ghostflyby.mill.project.MillDiscoveredModule
+import dev.ghostflyby.mill.settings.MillExecutionSettings
 
 internal object MillScalaSdkResolver {
     fun resolve(