+++ plugins/Mill/src/main/kotlin/dev/ghostflyby/mill/settings/MillLocalSettings.kt
@@ -20,29 +20,30 @@
  * <https://www.gnu.org/licenses/>.
  */
 
-package dev.ghostflyby.mill
+package dev.ghostflyby.mill.settings
 
-import com.intellij.openapi.externalSystem.service.settings.AbstractExternalSystemConfigurable
-import com.intellij.openapi.externalSystem.util.ExternalSystemSettingsControl
+import com.intellij.openapi.components.*
+import com.intellij.openapi.externalSystem.settings.AbstractExternalSystemLocalSettings
 import com.intellij.openapi.project.Project
+import dev.ghostflyby.mill.MillConstants
 
-internal class MillConfigurable(project: Project) :
-    AbstractExternalSystemConfigurable<MillProjectSettings, MillSettingsListener, MillSettings>(
-        project,
-        MillConstants.systemId,
-    ) {
-    override fun createProjectSettingsControl(settings: MillProjectSettings): ExternalSystemSettingsControl<MillProjectSettings> {
-        return MillProjectSettingsControl(settings)
+@Service(Service.Level.PROJECT)
+@State(name = "MillLocalSettings", storages = [Storage(StoragePathMacros.CACHE_FILE)])
+internal class MillLocalSettings(
+    project: Project,
+) : AbstractExternalSystemLocalSettings<MillLocalSettings.State>(MillConstants.systemId, project, State()),
+    PersistentStateComponent<MillLocalSettings.State> {
+
+    override fun getState(): State = state
+
+    override fun loadState(state: State) {
+        super.loadState(state)
     }
 
-    override fun createSystemSettingsControl(settings: MillSettings): ExternalSystemSettingsControl<MillSettings>? =
-        null
-
-    override fun newProjectSettings(): MillProjectSettings = MillProjectSettings()
-
-    override fun getId(): String = ID
+    class State : AbstractExternalSystemLocalSettings.State()
 
     companion object {
-        internal const val ID: String = "reference.settingsdialog.project.mill"
+        @JvmStatic
+        fun getInstance(project: Project): MillLocalSettings = project.service()
     }
 }